# internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go REST API
  description: |
    Production-ready REST API with JWT authentication, RBAC, and email verification.

    ## Authentication
    Most protected endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain a token via `POST /api/v1/auth/login`.
  version: "1.0"
  contact: {}
servers:
  - url: http://api.localhost
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateChannelRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: My Channel
        description:
          type: string
          maxLength: 500
          example: My channel description
    ChannelResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        avatar_url:
          type: string
        banner_url:
          type: string
        subscribers_count:
          type: integer
          format: int64
        videos_count:
          type: integer
          format: int64
        is_subscribed:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: validation failed
    ValidationFieldError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: must be a valid email address
    ValidationErrorDetail:
      allOf:
        - $ref: '#/components/schemas/ErrorDetail'
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: '#/components/schemas/ValidationFieldError'

    # ── Auth ────────────────────────────────────────────────────────────────────
...
  responses:
...
    401Unauthorized:
      description: Missing or invalid credentials / token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ErrorDetail'
          example:
            error:
              code: INVALID_CREDENTIALS
              message: invalid email or password
...
    409Conflict:
      description: Conflict — resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ErrorDetail'
          example:
            error:
              code: USER_ALREADY_EXISTS
              message: user already exists
    422Validation:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ValidationErrorDetail'
          example:
            error:
              code: VALIDATION_ERROR
              message: validation failed
              details:
                - field: email
                  message: must be a valid email address
                - field: username
                  message: must contain only letters, digits, dots, dashes, or underscores
...
    500Internal:
      description: Unexpected server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ErrorDetail'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: internal server error

# ────────────────────────────────────────────────────────────────────────────────
# PATHS
# ────────────────────────────────────────────────────────────────────────────────
paths:
  /api/v1/channels:
    post:
      tags: [channels]
      summary: Create the current user's channel
      operationId: createChannel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ChannelResponse'
        "401": { $ref: '#/components/responses/401Unauthorized' }
        "409": { $ref: '#/components/responses/409Conflict' }
        "422": { $ref: '#/components/responses/422Validation' }
        "500": { $ref: '#/components/responses/500Internal' }
...